Move local indexer to under indexer
