+++ .gitignore
@@ -6,3 +6,4 @@ redis/
 node_modules/
 .vscode/
 runner/yarn.lock
+**/.DS_Store
