+++ runner/src/indexer/local-indexer/local-indexer.ts
@@ -1,10 +1,10 @@
 import ContextBuilder, { type ContextObject } from '../context-builder';
-import InMemoryDmlHandler from '../dml-handler/in-memory-dml-handler';
-import IndexerConfig from '../indexer-config';
-import { type LocalIndexerConfig } from '../indexer-config/indexer-config';
-import NoOpIndexerMeta from '../indexer-meta/no-op-indexer-meta';
-import Indexer from '../indexer/indexer';
-import LakeClient from '../lake-client/lake-client';
+import InMemoryDmlHandler from '../../dml-handler/in-memory-dml-handler';
+import IndexerConfig from '../../indexer-config';
+import { type LocalIndexerConfig } from '../../indexer-config/indexer-config';
+import NoOpIndexerMeta from '../../indexer-meta/no-op-indexer-meta';
+import Indexer from '../../indexer/indexer';
+import LakeClient from '../../lake-client/lake-client';
 
 export default class LocalIndexer {
   public readonly indexer: Indexer;
